pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudlens"
version = "1.0.0"
description = "AI-powered multi-cloud network intelligence platform for SREs"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "jinja2>=3.1,<4",
    "pyyaml>=6,<7",
    "pydantic-settings>=2,<3",
    "prometheus-fastapi-instrumentator>=7,<8",
    "slowapi>=0.1.9,<1",
    "aiosqlite>=0.20,<1",
    "anthropic>=0.39,<1",
]

[project.optional-dependencies]
aws = ["boto3>=1.34,<2"]
azure = [
    "azure-identity>=1.17,<2",
    "azure-mgmt-resourcegraph>=8,<9",
]
gcp = ["google-cloud-compute>=1.19,<2"]
all-providers = ["cloudlens[aws,azure,gcp]"]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5,<6",
    "ruff>=0.4,<1",
    "mypy>=1.10,<2",
    "types-PyYAML>=6,<7",
    "httpx>=0.27,<1",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["config*", "providers*", "graph*", "db*", "engine*", "ai*", "api*", "exporters*"]
